Disables markers entering on zoom > 14.
If the client taps on a marker on zoom < 14 then the map will zoom to
the marker. This combats the crash on CoreJavascript created by trying
to zoom the map and going to a second state.

It also disables map interactions unless the marker returns `true` on
the `on-marker-tap` function.